requirements.txt
Django==4.2.16
djangorestframework==3.15.2
snowflake-connector-python==3.12.2
python-decouple==3.8